ai/text/hebImg2txt/workshop/requirements.txt
pillow
opencv-python
pytesseract
